+++ parts/header.php
@@ -30,7 +30,7 @@ $args = wp_parse_args($args, $array_defaults);
 				<?php get_template_part( 'parts/menu-bar') ?>
 			</div>
 			<!-- MENU CHEVRON BUTTON -->
-			<div class="site-menu-toggle">
+			<div id="site-menu-toggle" class="site-menu-toggle">
 				<label class="menu-label <?php if (is_search()) echo 'hidden'; else echo '' ?>">MENU</label>
 				<img id="menu-button" class="<?php if (is_search()) echo 'rotated'; else echo ''; ?>" src="<?php echo esc_url(get_template_directory_uri() . '/assets/images/chevron-down.svg'); ?>" title="<?php echo bloginfo('description') ?>" alt="<?php echo bloginfo('name'); ?>" width="32" height="auto" style="cursor: pointer">
 				<label class="menu-label <?php if (is_search()) echo ''; else echo 'hidden' ?>">MENU</label>