+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- `parts/header.php` now has the option to hide the menu
+- `page-login.php` and `page-registrazione.php` has been updated to use the new
+option in `parts-header.php` to make the pages more consistent.
+
 ## [0.5.0] 2025-08-06
 ### Added
 - Add Landing Page
@@ -72,7 +78,7 @@ the slug's name
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[unreleased]: https://github.com/erremauro/cigno-zen/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v0.5.0...HEAD
 [0.5.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.5.0
 [0.4.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.4.0
 [0.3.0]: https://github.com/erremauro/cigno-zen/releases/tag/v0.3.0